Move POI deserialization from url to util libs
